fix: many bugs

+++ src/main/java/com/dorm/controller/notice/NoticeController.java
@@ -1,13 +1,21 @@
 package com.dorm.controller.notice;
 
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.dorm.entity.notice.AddNoticeDTO;
 import com.dorm.entity.notice.NoticePO;
 import com.dorm.entity.notice.NoticeVO;
 import com.dorm.entity.user.UserPO;
+import com.dorm.entity.user.UserVO;
+import com.dorm.entity.user.teacher.TeacherPO;
+import com.dorm.enums.user.UserRoles;
+import com.dorm.enums.user.teacher.TeacherType;
 import com.dorm.service.notice.NoticeService;
 import com.dorm.service.user.UserService;
+import com.dorm.service.user.teacher.TeacherService;
+import com.dorm.utils.SecurityUtils;
 import jakarta.annotation.Resource;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,6 +29,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 public class NoticeController {
@@ -29,6 +38,12 @@ public class NoticeController {
 
     @Resource
     private UserService userService;
+
+    @Resource
+    private SecurityUtils securityUtils;
+    
+    @Resource
+    private TeacherService teacherService;
 
     @RequestMapping("/notice/list")
     public String listNotice(Model model) {
@@ -44,11 +59,11 @@ public class NoticeController {
     }
 
     @RequestMapping("/api/notice/add")
-    @PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER')")
     public String addNotice(
-            @ModelAttribute @Validated AddNoticeDTO noticeDTO,
-            BindingResult bindingResult,
-            RedirectAttributes redirectAttributes
+        @ModelAttribute @Validated AddNoticeDTO noticeDTO,
+        BindingResult bindingResult,
+        RedirectAttributes redirectAttributes
     ) {
         String url = "redirect:/notice/list";
 
@@ -58,14 +73,20 @@ public class NoticeController {
             return url;
         }
 
-        // 校验用户是否存在
-        UserPO userPO = userService.getById(noticeDTO.getUserId());
-        if (userPO == null) {
-            redirectAttributes.addFlashAttribute("msg", "用户不存在");
-            return url;
+        // 获取当前用户
+        UserVO currentUser = securityUtils.getCurrentUser();
+
+        // 如果是教师，判断是不是学办和学工部
+        if (currentUser.getRole() == UserRoles.TEACHER) {
+            TeacherPO teacherPO = teacherService.getOne(new QueryWrapper<TeacherPO>().eq("user_id", currentUser.getId()));
+            if (teacherPO.getTeacherType() != TeacherType.OFFICE && teacherPO.getTeacherType() != TeacherType.AFFAIRS) {
+                redirectAttributes.addFlashAttribute("msg", "只有管理员、学办和学工部可以发布公告");
+                return url;
+            }
         }
 
         NoticePO noticePO = NoticePO.valueOf(noticeDTO);
+        noticePO.setUserId(currentUser.getId());
         noticePO.setCreateTime(new Date());
         noticeService.save(noticePO);
 
@@ -73,9 +94,23 @@ public class NoticeController {
     }
 
     @RequestMapping("/api/notice/delete/{id}")
-    @PreAuthorize("hasRole('ADMIN')")
-    public String deleteNotice(@PathVariable Integer id) {
+    @PreAuthorize("hasAnyRole('ADMIN', 'TEACHER')")
+    public String deleteNotice(@PathVariable Integer id, RedirectAttributes redirectAttributes) {
+        String url = "redirect:/notice/list";
+
+        NoticePO noticePO = noticeService.getById(id);
+
+        // 获取当前用户
+        UserVO currentUser = securityUtils.getCurrentUser();
+
+        // 如果是教师，判断是不是他本人发布的
+        if (currentUser.getRole() == UserRoles.TEACHER && !Objects.equals(noticePO.getUserId(), currentUser.getId())) {
+            redirectAttributes.addFlashAttribute("msg", "学办和学工部只可以删除自己发布的公告");
+            return url;
+        }
+
         noticeService.removeById(id);
-        return "redirect:/notice/list";
+
+        return url;
     }
 }
